requirements.txt
streamlit>=1.32
pandas>=2.1
requests>=2.31
python-dotenv>=1.0
numpy>=1.26
python-dateutil>=2.8
bcrypt>=4.1
altair>=5.1
